# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "これはGo言語で作成されたOCRテスト用のシンプルなAPIです",
        "title": "OCR Tester API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/hello": {
            "get": {
                "description": "シンプルなGETリクエストで起動確認用のメッセージを返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Hello Worldメッセージを返す",
                "responses": {
                    "200": {
                        "description": "成功時のメッセージ",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/ocr-result/{id}": {
            "get": {
                "description": "Firestoreに保存された特定のOCR結果を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "OCR結果をIDで取得する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OCR結果のドキュメントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取得成功",
                        "schema": {
                            "$ref": "#/definitions/domain.OCRResult"
                        }
                    },
                    "404": {
                        "description": "指定されたIDが見つからない",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.OCRResult": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "extractedText": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "description": "FirestoreのドキュメントID",
                    "type": "string"
                }
            }
        }
    }
}